requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
easyocr>=1.7.1
opencv-python-headless>=4.13.0
Pillow>=12.2.0
numpy>=2.4.4
reportlab>=4.2.0
requests>=2.31.0
beautifulsoup4>=4.12.3
